Deleted LikeController.kt, move like-related logic to PostController.
Added logic to like a post and remove like.
