+++ src/main/kotlin/my/project/petsnap/dto/PostResponseDTO.kt
@@ -1,7 +1,5 @@
 package my.project.petsnap.dto
 
-import my.project.petsnap.entity.CommentDB
-import my.project.petsnap.entity.LikeDB
 import java.time.LocalDateTime
 
 data class PostResponseDTO(
@@ -9,6 +7,7 @@ data class PostResponseDTO(
     val image: String,
     val text: String?,
     val postTime: LocalDateTime,
-    val comments: List<CommentDB>?,
-    val likes: List<LikeDB>?
+    val commentsCount: Int,
+    val likesCount: Int,
+    val likedByUser: Boolean,
 )
