+++ src/main/kotlin/my/project/petsnap/dto/PostsOnMainPageResponseDTO.kt
@@ -1,7 +1,5 @@
 package my.project.petsnap.dto
 
-import my.project.petsnap.entity.CommentDB
-import my.project.petsnap.entity.LikeDB
 import java.time.LocalDateTime
 
 data class PostsOnMainPageResponseDTO(
@@ -10,6 +8,7 @@ data class PostsOnMainPageResponseDTO(
     val text: String?,
     val user: UserSearchResponseDTO,
     val postTime: LocalDateTime,
-    val comments: List<CommentDB>?,
-    val likes: List<LikeDB>?
+    val commentsCount: Int,
+    val likesCount: Int,
+    val likedByUser: Boolean,
 )
