+++ src/main/kotlin/my/project/petsnap/entity/LikeDB.kt
@@ -15,4 +15,8 @@ data class LikeDB(
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "post_id")
     var post: PostDB,
+
+//    @ManyToOne(fetch = FetchType.LAZY)
+//    @JoinColumn(name = "comment_id")
+//    var comment: CommentDB?
 )
